add single InformingLayout
