+++ app/src/main/AndroidManifest.xml
@@ -33,7 +33,7 @@
         <receiver
             android:name=".receiver.IncomingCallReceiver"
             android:enabled="true"
-            android:exported="false">
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.PHONE_STATE" />
             </intent-filter>
