+++ app/src/main/java/ru/yandex/whocallsya/receiver/IncomingCallReceiver.java
@@ -3,6 +3,7 @@ package ru.yandex.whocallsya.receiver;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
+import android.util.Log;
 
 import ru.yandex.whocallsya.service.CockyBubblesService;
 
@@ -15,9 +16,11 @@ public class IncomingCallReceiver extends BroadcastReceiver {
 
     @Override
     public void onReceive(Context context, Intent intent) {
+        Log.e("whocallsya","IncomingCallReceiver");
         if (!intent.getStringExtra(EXTRA_STATE).equals(EXTRA_STATE_RINGING)) return;
         Intent i = new Intent(context, CockyBubblesService.class);
         i.putExtra(PHONE_NUMBER, intent.getStringExtra(EXTRA_INCOMING_NUMBER));
+        Log.e("whocallsya","IncomingCallReceiver" + intent.getStringExtra(EXTRA_INCOMING_NUMBER));
         context.startService(i);
     }
 }
